+++ model/src/main/java/com/viclab/model/repository/Repository.kt
@@ -2,7 +2,7 @@ package com.viclab.model.repository
 
 data class Repository(
     val name: String,
-    val score: Float,
+    val score: Int,
     val forks: Int,
     val owner: Owner,
 )
